# apps/disher-backend-3.0/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Disher API",
    "version": "1.0.0",
    "description": "Backend for the Disher food diary. Errors share one RFC 9457-subset problem+json shape — see api/errors.ts."
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-disher1.session_token"
      }
    },
...
  },
  "paths": {
...
    "/articles/nutrients/{folder}": {
      "get": {
        "operationId": "getNutrientArticle",
        "tags": [
          "content"
        ],
        "description": "Raw markdown of one nutrient article.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "folder",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/matcher-telemetry/": {
      "post": {
        "operationId": "logMatcherTelemetry",
        "tags": [
          "telemetry"
        ],
        "description": "Anonymous matcher-quality beacon. The ONE route exempt from the trusted-origin guard — sendBeacon carries neither Origin nor credentials. Carries no session and mutates no user data.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "title": "MatcherTelemetryEvent",
                "type": "object",
                "required": [
                  "requestId",
                  "userId",
                  "action",
                  "itemsTotal",
                  "itemsCommitted",
                  "itemsDeleted",
                  "itemsWithEditedFood",
                  "itemsWithEditedTime",
                  "itemsWithEditedQty",
                  "llmLatencyMs",
                  "matcherLatencyMs",
                  "reviewDurationMs",
                  "corrections"
                ],
                "properties": {
                  "requestId": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "action": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "commit"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "abandon"
                        ]
                      }
                    ]
                  },
                  "itemsTotal": {
                    "type": "number"
                  },
                  "itemsCommitted": {
                    "type": "number"
                  },
                  "itemsDeleted": {
                    "type": "number"
                  },
                  "itemsWithEditedFood": {
                    "type": "number"
                  },
                  "itemsWithEditedTime": {
                    "type": "number"
                  },
                  "itemsWithEditedQty": {
                    "type": "number"
                  },
                  "llmLatencyMs": {
                    "type": "number"
                  },
                  "matcherLatencyMs": {
                    "type": "number"
                  },
                  "reviewDurationMs": {
                    "type": "number"
                  },
                  "corrections": {
                    "type": "array",
                    "items": {
                      "additionalProperties": false,
                      "title": "MatcherCorrection",
                      "type": "object",
                      "required": [
                        "originalName",
                        "matcherChoice",
                        "userChoice",
                        "correctionType"
                      ],
                      "properties": {
                        "originalName": {
                          "type": "string"
                        },
                        "matcherChoice": {
                          "type": "string"
                        },
                        "userChoice": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "correctionType": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "accepted-top1"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "switched-ambiguous"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "manual-search"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "deleted"
                              ]
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/backup/": {
      "put": {
        "operationId": "putBackup",
        "tags": [
          "backup"
        ],
        "description": "Write the user's snapshot. Last write wins.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "BackupSnapshot",
                "description": "Opaque client snapshot (every Dexie table). The server stores it as jsonb and never inspects it."
              }
            }
          },
          "description": "Opaque client snapshot (every Dexie table). The server stores it as jsonb and never inspects it."
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "operationId": "getBackup",
        "tags": [
          "backup"
        ],
        "description": "Read the user's snapshot, or 404 when none exists.",
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "operationId": "deleteBackup",
        "tags": [
          "backup"
        ],
        "description": "Erase the user's server-side vault. Idempotent — 204 either way.",
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/user-reports/": {
      "post": {
        "operationId": "submitUserReport",
        "tags": [
          "feedback"
        ],
        "description": "User-facing «Сообщить о проблеме» sink — text + client metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "title": "UserReport",
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "page": {
                    "type": "string"
                  },
                  "screenSize": {
                    "type": "string"
                  },
                  "userAgent": {
                    "type": "string"
                  },
                  "pwa": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/admin/user-reports": {
      "get": {
        "operationId": "listUserReports",
        "tags": [
          "admin"
        ],
        "description": "Production bug reports (table user_reports) joined with the author's email, newest first. `q` matches email / text / user id.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Clamped to 1..500 by the handler; default 100."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          }
        ],
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}